backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.7
langchain-experimental==0.0.61
langchain-groq==0.1.4
groq==0.9.0
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
python-dotenv==1.0.1
pydantic==2.7.4